Organize api by resource
