+++ src/views/ads/Report.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { useParams, useNavigate, useSearchParams } from 'react-router-dom'
 import { Container, Title, Paper, Stack, Group, Text, Loader, Breadcrumbs, Anchor } from '@mantine/core'
 import { IconChevronRight } from '@tabler/icons-react'
-import api from '@/utils/api'
+import api from '@/api'
 import DataTable from '@/components/DataTable'
 import { useAsync } from '@/hooks/useAsync'
 import classes from '@/styles/AdsReport.module.css'
